backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
redis==5.0.1
celery==5.3.6
httpx==0.26.0
reportlab==4.0.9
openpyxl==3.1.2
minio==7.2.3
python-dateutil==2.8.2
